Added query_games_info() in DBreader.py Added load_games method in FilesLoader.py
Added Controller.py and add_game().

Added all the handlers and methods relative to /myPCs and /createGame in Bot.py.

Added in ENG.json all the phrases accordingly to new methods added in Bot.py.

Changed the Game.py constructor: now the default for the identifier and the title are None, and they are set inside the method.
Added DocStrings to the Game __init__